migration

+++ Web-eCommerce/database/migrations/2020_02_24_105541_create_product_value_pivot_table.php
@@ -13,9 +13,9 @@ class CreateProductValuePivotTable extends Migration
     public function up()
     {
         Schema::create('product_value', function (Blueprint $table) {
-            $table->integer('product_id')->unsigned()->index();
+            $table->unsignedBigInteger('product_id')->unsigned()->index();
             $table->foreign('product_id')->references('id')->on('products')->onDelete('cascade');
-            $table->integer('value_id')->unsigned()->index();
+            $table->unsignedBigInteger('value_id')->unsigned()->index();
             $table->foreign('value_id')->references('id')->on('values')->onDelete('cascade');
             $table->primary(['product_id', 'value_id']);
         });
